# Consistently return StringValues.Empty instead of default for missing headers