--- a/src/Http/Http/test/HeaderDictionaryTests.cs
+++ b/src/Http/Http/test/HeaderDictionaryTests.cs
@@ -113,28 +113,23 @@ public class HeaderDictionaryTests
         Assert.Equal(new[] { "value " }, result);
     }
 
-    [Fact]
-    public void PublicIndexerReturnsStringValuesEmptyForMissingHeaders()
+    [Theory]
+    [InlineData(true)]
+    [InlineData(false)]
+    public void ReturnsCorrectStringValuesEmptyForMissingHeaders(bool withStore)
     {
-        var headers = new HeaderDictionary();
+        // Test both with and without HeaderDictionary.Store set.
+        var emptyHeaders = withStore ? new HeaderDictionary(1) : new HeaderDictionary();
 
         // StringValues.Empty.Equals(default(StringValues)), so we check if the implicit conversion
         // to string[] returns null or Array.Empty<string>() to tell the difference.
-        Assert.Same(Array.Empty<string>(), (string[])headers["Header1"]);
-    }
+        Assert.Same(Array.Empty<string>(), (string[])emptyHeaders["Header1"]);
 
-    [Fact]
-    public void IHeaderDictionaryMembersReturnStringValuesEmptyForMissingHeaders()
-    {
-        IHeaderDictionary headers = new HeaderDictionary();
-        Assert.Same(Array.Empty<string>(), (string[])headers["Header1"]);
-        Assert.Same(Array.Empty<string>(), (string[])headers.Host);
-    }
+        IHeaderDictionary asIHeaderDictionary = emptyHeaders;
+        Assert.Same(Array.Empty<string>(), (string[])asIHeaderDictionary["Header1"]);
+        Assert.Same(Array.Empty<string>(), (string[])asIHeaderDictionary.Host);
 
-    [Fact]
-    public void IDictionaryIndexerThrowsForMissingHeaders()
-    {
-        IDictionary<string, StringValues> headers = new HeaderDictionary();
-        Assert.Throws<KeyNotFoundException>(() => headers["Header1"]);
+        IDictionary<string, StringValues> asIDictionary = emptyHeaders;
+        Assert.Throws<KeyNotFoundException>(() => asIDictionary["Header1"]);
     }
 }