Fixes #44509