@JamesNK 2023-01-24
Could there be any performance impact here? There is lots of setting, getting and resetting of headers per request.